+++ NEWS.md
@@ -29,6 +29,7 @@
 - Update vocabulary term IRIs in tests to replace 'marc' with 'library' [MODLD-792](https://folio-org.atlassian.net/browse/MODLD-792)
 - Map MARC 240 to HUB resource [MODLD-884](https://folio-org.atlassian.net/browse/MODLD-884)
 - Isolated Marc4LdObjectMapper introduced [MODLD-873](https://folio-org.atlassian.net/browse/MODLD-873)
+- Map MARC 130 to HUB resource [MODLD-879](https://folio-org.atlassian.net/browse/MODLD-879)
 
 ## 1.0.3 (04-09-2025)
 - Geographic coverage: single "geographicCoverage" edge is created for Work which contains 043 with multiple $a fields [MODLD-694](https://folio-org.atlassian.net/browse/MODLD-694)